automation/requirements.txt
python-dotenv
pymongo
matplotlib
reportlab
